% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\docType{data}
\name{swamids}
\alias{swamids}
\title{SWAM Model Object IDs joined to Intake SourceIDs.}
\format{
A dataframe with 4 columns and over 500 rows.
}
\usage{
swamids
}
\description{
Each water use object in the SWAM models can be represented as a set of intakes.
\describe{
\item{\code{swamid}} {SWAM model object ID representing a given water user.}
\item{\code{sourceid}} {An intake sourceid associated with the swamid}
\item{\code{swam_model}} {which SWAM basin model the data is from.}
\item{\code{comments}}{}
}
}
\examples{
data(swamids)
head(swamids)
}
\keyword{datasets}
